refactor(categorycards): rename to category_card.dart
