+++ lib/views/components/category_card.dart
