+++ lib/views/socialize.dart
@@ -1,7 +1,7 @@
 import '../core/category.dart';
 import '../providers/application_provider.dart';
 import 'components/app_list.dart';
-import 'components/category_cards.dart';
+import 'components/category_card.dart';
 import 'components/theme.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
